Split cacheWatcher into its own file
